Add ConnectionProxy class for DB Connection management
Schemaman uses the `__del__` methods of various objects to handle closing/releasing resources. In the current implementation we actually are creating a bunch of self-refferential objects.

Specifically:
    - connection -> request
    - DB_CONNECTION_POOL -> connection

This is problematic as the request deletion is what releases the connection-- so that it can be used by other requests. So, to break this dep chain I've added ConnectionProxy which is just a proxy object which points at the Connection object. When this ConnectionProxy is destroyed it removes the Connection's pointer at the request-- which in turn lets the request GC-- which will release the connection from the request.