Fix for mis-ordering of dependant table inserts
When a VersionRecord is being comitted to a datasource we need to do the inserts in a specific order such that we can update the various ids and adhere to foreign key constraints.

The high-level idea is to take the list of tables -- and pop them from the potential list into the ordered list once its table dependencies are no longer in the initial list of tables.